requirements.txt
Django==4.2.14         # LTS estable
Pillow==10.4.0
pandas==2.2.2
matplotlib==3.9.0
python-decouple==3.8
python-dotenv==1.0.1
openai==1.44.1
numpy==1.26.4
requests==2.32.3